+++ .github/workflows/pytest-tests.yml
@@ -44,7 +44,7 @@ jobs:
         flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
     - name: Generate Test Data
       run: |
-        python ./EAB_tools/tests/io/data/generate-all-test-data.py
+        python ./EAB_tools/tests/io/data/generate_all_test_data.py
     - name: Test with pytest
       run: >
         pytest
