+++ .github/workflows/quick-pytest-tests.yml
@@ -43,7 +43,7 @@ jobs:
         flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
     - name: Generate Test Data
       run: |
-        python ./EAB_tools/tests/io/data/generate-all-test-data.py
+        python ./EAB_tools/tests/io/data/generate_all_test_data.py
     - name: Quick test with pytest
       run: >
         pytest
